+++ log/__init__.py
@@ -9,6 +9,8 @@ d = debug
 i = info
 w = warn = warning
 e = error
+c = critical
+exc = exception
 
 __project__ = 'minilog'
-__version__ = '0.3.1'
+__version__ = '0.4'