+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision History
 
+## 0.4 (2018/4/28)
+
+- Added `reset=True` as `init()` option to replace all existing logging handlers.
+- Added `exception` logging API.
+- Added convenience alias: `log.c`, `log.exc`.
+
 ## 0.3.1 (2018/03/30)
 
 - Fixed bug where records were written for disabled levels.